Name refactorings
