+++ test/express/integration/tables/configuration.tests.js
@@ -4,8 +4,8 @@
 var expect = require('chai').expect,
     supertest = require('supertest'),
     express = require('express'),
-    mobileApps = require('../../../infrastructure/mobileApps').ignoreEnv,
-    config = require('../../../infrastructure/config'),
+    mobileApps = require('../../../infrastructure/appFactory').ignoreEnvironment,
+    config = require('../../../infrastructure/configuration'),
 
     app, mobileApp;
